# [Torch] Restore class-aware NMS for detection models by graph rewrite