@zhiics 2020-12-23
@masahi Thanks for the perf improvement. Could you provide the CPU numbers as well?