@masahi 2020-12-23
@zhiics Sure updated the description. The regression is only 200 us on CPU, so it may be just a measurement noise.

I have no idea why I'm not getting good speed up. IOU tests, including memory access to boxes should be definitely reduced. The only additional overhead I think of is that the input to NMS is one column wider, due to storing class ids. 

Performance is not great, but I believe having access to class ids should not be bad idea...