@masahi 2020-12-23
I should mention that this rewrite is not run by default, so there is no perf risk.